Moving JS into different file
